removed UserRole and put all permissions directly on tournament.  added more stuff to tournament models

+++ server/controllers/tournaments.js
@@ -1,47 +1,35 @@
 const Tournament = require('../models/tournament')
-const UserRole = require('../models/userRole')
 const moment = require('moment')
 const logger = require('../logging/logger')
 
 const getTournamentsHandler = [
   async function getTournamentsFunc(req, res) {
-    let isActive = req.params.isActive
-    if (isActive === undefined) isActive = false
-    const tournaments = await Tournament.find({ isActive: isActive }).exec()
+    const tournaments = await Tournament.find().exec()
     res.send(tournaments)
   }
 ]
 
 const createTournamentHandler = [
   async function createTournamentFunc(req, res) {
-    const userId = res.locals.user._id
-    let savedTournament
     try {
       logger.info('creating tournament', req.body)
+      const userId = res.locals.user._id
       const body = req.body
       const newT = new Tournament()
+
       newT.title = body.title
       newT.startDate = moment(body.startDate).toDate()
       newT.endDate = moment(body.endDate).toDate()
       newT.creator = res.locals.user._id
       newT.isActive = false
+      newT.organizers = [userId]
 
-      savedTournament = await newT.save()
+      const savedTournament = await newT.save()
+      res.send(savedTournament)
     } catch (e) {
       logger.error(e)
       res.status(500).send({ error: 'unable to create tournament' })
     }
-    try {
-       await UserRole.update(
-         {userId: userId, tournamentId: savedTournament._id},
-        {role: 'TO'},
-        {upsert: true})
-        .exec()
-    } catch (e) {
-      logger.warn(`unable to create TO role for ${userId}`)
-    }
-
-    res.send(savedTournament)
   }
 ]
 
@@ -52,19 +40,12 @@ const cancelTournamentHandler = [
       const userId = res.locals.user._id
       logger.info(`cancelling tournament ${tournamentId}`)
 
-      const isAllowedResult = await UserRole.count({tournamentId, userId, role: 'TO'}).exec()
-
-      if (isAllowedResult !== 1) {
-        logger.info(`${userId} not allowed to cancel tournament ${tournamentId}`)
-        return res.status(403).send()
-      }
-      const response = await Tournament.deleteOne({_id: tournamentId}).exec()
+      const response = await Tournament.deleteOne({ _id: tournamentId, organizers: userId }).exec()
       if (!response.result.ok) {
         logger.error(`${userId} unable to cancel tournament ${tournamentId}`)
         return res.status(404).send()
       }
-      await UserRole.deleteMany({tournamentId}).exec()
-      res.send({tournamentId, userId})
+      res.send({ tournamentId, userId })
     } catch (e) {
       logger.error(e)
       res.status(500).send({ error: 'unable to cancel tournament' })
@@ -80,14 +61,14 @@ const joinTournamentHandler = [
       logger.info(`${userId} joining tournament ${tournamentId}`)
 
       const result = await Tournament.update(
-        { _id: tournamentId, isActive: false, entrants: { $ne: userId } },
-        {$addToSet: {entrants: userId}})
+        { _id: tournamentId, isActive: false },
+        { $addToSet: { entrants: userId } })
         .exec()
       if (!result.ok) {
         logger.info(`${userId} unable to join tournament ${tournamentId}`)
         return res.status(404).send()
       }
-      res.send({tournamentId, userId})
+      res.send({ tournamentId, userId })
     } catch (e) {
       logger.error(e)
       res.status(500).send({ error: 'unable to join tournament' })
@@ -103,14 +84,14 @@ const leaveTournamentHandler = [
       logger.info(`${userId} leaving tournament ${tournamentId}`)
 
       const result = await Tournament.update(
-        { _id: tournamentId, isActive: false, entrants: { $ne: userId } },
-        {$pull: {entrants: userId}})
+        { _id: tournamentId, isActive: false },
+        { $pull: { entrants: userId } })
         .exec()
       if (!result.ok) {
         logger.info(`${userId} unable to leave tournament ${tournamentId}`)
-        return res.send({tournamentId, userId})
+        return res.send({ tournamentId, userId })
       }
-      res.send({tournamentId, userId})
+      res.send({ tournamentId, userId })
     } catch (e) {
       logger.error(e)
       res.status(500).send({ error: 'unable to leave tournament' })
@@ -125,9 +106,9 @@ const activateTournamentHandler = [
       const userId = res.locals.user._id
       logger.info(`${userId} activating tournament ${tournamentId}`)
       const tournament = await Tournament.findOneAndUpdate(
-        {_id: tournamentId, isActive: false},
-        {isActive: true},
-        {new: true})
+        { _id: tournamentId, isActive: false, organizers: userId },
+        { isActive: true },
+        { new: true })
         .exec()
       res.send(tournament)
     } catch (e) {
